--- test/integration/export/next.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  exportPathMap: () => ({
-    '/': { page: '/', query: { a: 'blue' } }
-  })
-}
