+++ test/integration/static/next.config.js
@@ -15,7 +15,8 @@ module.exports = (phase) => {
         '/dynamic': { page: '/dynamic', query: { text: 'cool dynamic text' } },
         '/dynamic/one': { page: '/dynamic', query: { text: 'next export is nice' } },
         '/dynamic/two': { page: '/dynamic', query: { text: 'zeit is awesome' } },
-        '/file-name.md': { page: '/dynamic', query: { text: 'this file has an extension' } }
+        '/file-name.md': { page: '/dynamic', query: { text: 'this file has an extension' } },
+        '/query': { page: '/query', query: { a: 'blue' } }
       }
     }
   }
