+++ test/integration/static/pages/query.js
